+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for azw3 (Amazon Kindle Format) documents.
 - Added the support of OpenType standard. It is usefull for languages required advanced typography.
+- Added 'MaxImageResolution' property for SvgSaveOptionsData class.
 
 
 ## [24.6.0] - Aspose Words Cloud for Dart 24.6 Release Notes